[Wasm GC] [ctor-eval] Evaluate and serialize GC data (#4491)
This ended up simpler than I thought. We can simply emit global and
local data as we go, creating globals as necessary to contain GC data,
and referring to them using global.get later. That will ensure that
data identity works (things referring to the same object in the interpreter
will refer to the same object when the wasm is loaded). In more detail,
each live GC item is created in a "defining global", a global that is
immutable and of the precise type of that data. Then we just read from
that location in any place that wants to refer to that data. That is,
something like

function foo() {
  var x = Bar(10);
  var y = Bar(20);
  var z = x;
  z.value++; // first object now contains 11
  ...
}

will be evalled into something like

var define$0 = Bar(11); // note the ++ has taken effect here
var define$1 = Bar(20);

function foo() {
  var x = define$0;
  var y = define$1;
  var z = define$0;
  ...
}

This PR should handle everything but "cycles", that is, GC data that at
runtime ends up forming a loop. Leaving that for later work (not sure
how urgent it is to fix).