[EH] Optimize out trys whose body does not throw (DO NOT LAND)
This performs an interprocedural analysis using the call graph to decide
whether each function can throw or not, and use the info to optimize out
trys whose body does not throw.

About half of the code is for printing information and not for actual
optimization.

I experimented with 'wasm-opt' wasm binary built with -O2, and while
this pass was able to analyze 6975/7968 (87.5%) functions for
throwability, this ended up optimizing out only 162/9790 (1.7%) trys,
resulting in a mere 0.02% improvement in code size. Not analyzable
functions were considered throwable.

The main reason was there are not very many functinons that can't throw.
Most functions call either imported library functions or C++ standard
library functions. Imported functions are considered throwable, and most
C++ library functions can throw. And many library functions also contain
indirect calls, which we consider throwable. Maybe we can try searching
all modules to figure out if there is any functions with the specific
signature that can throw to figure out whether a `call_indirect` throws
or not, but this pass doesn't do it yet. Also we can think of a way to
break more call graph cycles to make more functions analyzable, but I'm
doubt if it will bring meaningful improvement given the current stat.

All imported functions should be considered throwable. But as an
experiment, I also tried to measure the possible improvement in code
size when we assume all imported functions can't throw. Even if this
incorrectly optimistic experiment, the pass was able to optimize
412/9790 (4.2%) trys, resulting in 0.09% improvement.

I'm planning to run this on other binaries too, but at the moment, due
to the disappointing results, I'm not necessarily planning to land this.
But I'm uploading this anyway for future reference, in case I, or
someone else, gives another try at optimizing this.