[NFC] Move and generalize parameter-removing logic from DeadArgumentElimination (#4544)
In preparation for removing dead arguments from all functions sharing a heap
type (which seems useful for j2wasm output), first this PR refactors that code
so it is reusable. This moves the code out of the pass into FunctionUtils, and
also generalizes it slightly by

supporting a set of functions and not just a single one, and
receiving a list of call_refs and not just calls
(no other changes to anything).