[Wasm GC] Signature Pruning (#4545)
This adds a new signature-pruning pass that prunes parameters from
signature types where those parameters are never used in any function
that has that type. This is similar to DeadArgumentElimination but works
on a set of functions, and it can handle indirect calls.

Also move a little code from SignatureRefining into a shared place to
avoid duplication of logic to update signature types.

This pattern happens in j2wasm code, for example if all method functions
for some virtual method just return a constant and do not use the this
pointer.