Refined HeapTypes with depth bounds
Add first-class support for tracking additional information through the type
system in the form of refined heap types, i.e. heap types with additional
information that restricts the sets of values that can inhabit those types.
These refined types are implemented as a new kind of HeapType whose HeapTypeInfo
contains an unrefined base HeapType and a `Refinement` object containing the
additional information. For now that information is a "subtype depth" that
controls how deep in the subtype hierarchy the dynamic type of values inhabiting
the refined type can be. Tracking subtype depth is a generalization over
tracking whether or not types are "exact" that is still relatively cheap. We may
add additional information to `Refinement` objects in the future, for example to
differentiate between duplicate type definitions.

To avoid having to change all users of the type API to account for the
possibility that any HeapType may be refined, update most of the functions in
the API to look through any refinement.

Update subtyping and LUB calculation to take subtype depth refinements into
account.

A follow-on PR will start using refined types in the optimizer.