RemoveUnusedModuleElements: Track CallRef/RefFunc more precisely (#4621)
If we see (ref.func $foo) that does not mean that $foo is reachable - we
must also see a (call_ref ..) of the proper type. Only after seeing both should
we mark the function as reachable, which this PR does.

This adds some complexity as we need to track intermediate state as we go,
since we could see the RefFunc before the CallRef or vice versa. We also
need to handle the case of a RefFunc without a CallRef properly: We cannot
remove the function, as the RefFunc must refer to it, but at least we can
empty out the body since we know it is never reached.

This removes an old wasm-opt test which is now superseded by a new lit
test.

On J2Wasm output this removes 3% of all functions, which account for
2.5% of total code size.