Handle call.without.effects in RemoveUnusedModuleElements (#4624)
We assume a closed world atm in the GC space, but the call.without.effects
intrinsic sort of breaks that: that intrinsic looks like an import, but we really
need to care about what is sent to it even in a closed world:

(call $call-without-effects
  (ref.func $target-keep)
)

That reference cannot be ignored, as logically it is called just as if there
were a call_ref there. This adds support for that, fixing the combination of
#4621 and using call.without.effects.

Also flip the vector of ref.func names to a set. I realized that in a very
large program we might see the same name many times.