Remove renameMainArgcArgv from wasm-emscripten-finalize (#4700)
This part to finalize is currently not used and was added in preparation
for https://reviews.llvm.org/D75277.

However, the better solution to dealing with this alternative name for
main is on the emscripten side.  The main reason for this is that
doing the rename here in binaryen would require finalize to always
re-write the binary, which is expensive.