[Parser] Begin parsing modules (#4716)
Implement the basic infrastructure for the full WAT parser with just enough
detail to parse basic modules that contain only imported globals. Parsing
functions correspond to elements of the grammar in the text specification and
are templatized over context types that correspond to each phase of parsing.
Errors are explicitly propagated via `Result<T>` and `MaybeResult<T>` types.

Follow-on PRs will implement additional phases of parsing and parsing for new
elements in the grammar.