[WasmGC] OptimizeInstructions: Improve RefIs cast ordering (#4752)
#4748 regressed us in some cases, because it removed casts first:

  (ref.is_func
    (ref.as_func
      (local.get $anyref)))

If the cast is removed first, and the local has no useful type info, then
we'd have removed the cast but could not remove the ref.is. But
the ref.is could be optimized to 1, as it must be a func - the type
info proves it thanks to the cast. To avoid this, remove casts after
everything else.