Move method implementations to drop.cpp
Now that we have `getDroppedUnconditionalChildrenAndAppend`, all passes
need to use it rather than the old `getDroppedChildrenAndAppend`, which
is used within the new method
`getDroppedUnconditionalChildrenAndAppend`.

This creates `drop.cpp` and move method implementations in drop.h there,
leaving only the declaration of
`getDroppedUnconditionalChildrenAndAppend` in drop.h, making the old
`getDroppedChildrenAndAppend` inaccessible from the outside.

Also replace a usage of the old method in `OptimizeInstruction` to a new
one. Couldn't come up with a test case that shows a different output
between the old and the new method, because by the time we get to
`visitRefEq`, many things have been simplified already.