Simplify routines for dropping children (NFC)
I think we can actually move all the checking routines to
`getDroppedUnconditionalChildrenAndAppend` and remove `canRemove` and
`canRemoveStructurally`. Checking things in three different places was a
little hard to understand, and I think
`getDroppedUnconditionalChildrenAndAppend` actually needs to exclude
more things that `canRemoveStructurally` covers, if it is to be used in
other places as well.

Additionally, I think it'd better to make these two dropping-children
functions (`getDroppedChildrenAndAppend` and
`getDroppedUnconditionalChildrenAndAppend`) to a cpp file and make
`getDroppedChildrenAndAppend` an internal function, inaccessible to
outer passes, given that it looks all places should use
`getDroppedUnconditionalChildrenAndAppend` instead. But this can be a
follow-up.