[GUFA] Ignore trapping children when unreachable
In `getDroppedChildrenAndAppend`, when the type of the given last
expression is unreachable, we ignore trapping effects when deciding
whether to drop the current expression because the last expression will
trap anyway.

We can do the same thing when dropping children. This will make GUFA's
generated code a little simpler in some cases. In GUFA's case these are
supposed to be optimized away by DCE anyway, but I think it's good to
handle this in this utility function too.