Change stack pass to use existing global variables.
With this change, toolchains can specify `__stack_base` and `__stack_end`
global variables which the binaryen stack pass uses directly to check against
overflows, instead of the pass emitting its own variables and setter (which
previously was `__set_stack_limits`, and toolchains had to invoke this setter).

If the globals are not declared, the stack pass is skipped.